gbsdfbds

+++ tagm/tagm/core/pipeline.py
@@ -79,7 +79,7 @@ class Pipeline:
         log("loading", f"Loading instruct model: {self.instruct_model_id}")
         self.instruct_model = AutoModelForCausalLM.from_pretrained(
             self.instruct_model_id,
-            torch_dtype=self.dtype,
+            dtype=self.dtype,
             device_map=self.device,
             attn_implementation="eager",
             token=self.hf_token,
@@ -129,7 +129,7 @@ class Pipeline:
         log("loading", f"Loading base model: {self.base_model_id}")
         self.base_model = AutoModelForCausalLM.from_pretrained(
             self.base_model_id,
-            torch_dtype=self.dtype,
+            dtype=self.dtype,
             device_map=self.device,
             attn_implementation="eager",
             token=self.hf_token,
